[services] Move memory allocator classes into a new module for testability
